requirements.txt
Django==2.2.7
psycopg2
Pillow
autopep8
django-phonenumber-field
phonenumbers
